% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vcf.R
\name{get_vcf}
\alias{get_vcf}
\title{Get data for a genomic region from a remote VCF file.}
\usage{
get_vcf(chrom, start, end, pop = NA)
}
\arguments{
\item{chrom}{a chromosome name (1-22,X) without "chr"}

\item{start}{a positive integer indicating the start of a genomic region}

\item{end}{a positive integer indicating the end of a genomic region}

\item{pop}{the name of a 1000 Genomes population (AMR,AFR,ASN,EUR,...)}
}
\value{
A list with three dataframes:
\describe{
  \item{ind}{A dataframe with information about individuals: Family.ID,
   Individual.ID, Paternal.ID, Maternal.ID, Gender, Population,
   Relationship, Siblings, Second.Order, Third.Order, Other.Comments,
   SuperPopulation}
  \item{meta}{First 8 columns of the VCF file: CHROM, POS, ID, REF, ALT,
   QUAL, FILTER, INFO}
  \item{geno}{Columns 10 onward of the VCF file. All genotypes are converted
   to 0s and 1s representing REF and ALT alleles. This dataframe has two
   columns for each individual.}
}
}
\description{
Returns a list with three dataframes for individuals, SNPs, and genotypes.
}
\details{
Currently, this is hard-coded to access 1000 Genomes phase3 data hosted by
Brian Browning (author of BEAGLE):

\url{http://bochet.gcc.biostat.washington.edu/beagle/1000_Genomes_phase3_v5a/}

This implementation discards multi-allelic markers that have a "," in the
ALT column.

The \code{pop} can be any of: ACB, ASW, BEB, CDX, CEU, CHB, CHS, CLM, ESN,
FIN, GBR, GIH, GWD, IBS, ITU, JPT, KHV, LWK, MSL, MXL, PEL, PJL, PUR, STU,
TSI, YRI. It can also be any super-population: AFR, AMR, EAS, EUR, SAS.

Find more details here:
\url{http://www.1000genomes.org/faq/which-populations-are-part-your-study}
}
\examples{
vcf <- get_vcf(chrom = "12", start = 533090, end = 623090, pop = "AFR")
names(vcf)
 
}

